fix: Use onClose handler for notification
