+++ src/components/common/Notifications/index.tsx
@@ -6,16 +6,12 @@ import type { Notification } from '@/store/notificationsSlice'
 import { closeNotification, readNotification, selectNotifications } from '@/store/notificationsSlice'
 import type { AlertColor, SnackbarCloseReason } from '@mui/material'
 import { Alert, Link, Snackbar } from '@mui/material'
-import { type SafeInfo } from '@safe-global/safe-gateway-typescript-sdk'
 import css from './styles.module.css'
 import NextLink from 'next/link'
 import ChevronRightIcon from '@mui/icons-material/ChevronRight'
 import { OVERVIEW_EVENTS } from '@/services/analytics/events/overview'
 import Track from '../Track'
 import { isRelativeUrl } from '@/utils/url'
-import useSafeInfo from '@/hooks/useSafeInfo'
-import { DISMISS_NOTIFICATION_KEY, isUpdateSafeNotification } from '@/hooks/useSafeNotifications'
-import useLocalStorage from '@/services/local-storage/useLocalStorage'
 
 const toastStyle = { position: 'static', margin: 1 }
 
@@ -101,44 +97,18 @@ export type DismissedUpdateNotifications = {
   }
 }
 
-const handleDismissUpdateSafeNotification = (
-  notification: Notification,
-  safe: SafeInfo,
-  dismissedNotifications: ReturnType<typeof useLocalStorage<DismissedUpdateNotifications>>[0],
-  setDismissedNotifications: ReturnType<typeof useLocalStorage<DismissedUpdateNotifications>>[1],
-) => {
-  const EXPIRY_DAYS = 90
-
-  if (!isUpdateSafeNotification(notification.groupKey)) return
-
-  const expiryDate = Date.now() + EXPIRY_DAYS * 24 * 60 * 60 * 1000
-
-  const newState = {
-    ...dismissedNotifications,
-    [safe.chainId]: {
-      ...dismissedNotifications?.[safe.chainId],
-      [safe.address.value]: expiryDate,
-    },
-  }
-
-  setDismissedNotifications(newState)
-}
-
 const Notifications = (): ReactElement | null => {
-  const [dismissedNotifications, setDismissedNotifications] =
-    useLocalStorage<DismissedUpdateNotifications>(DISMISS_NOTIFICATION_KEY)
   const notifications = useAppSelector(selectNotifications)
   const dispatch = useAppDispatch()
-  const { safe } = useSafeInfo()
 
   const visible = getVisibleNotifications(notifications)
 
   const handleClose = useCallback(
     (item: Notification) => {
       dispatch(closeNotification(item))
-      handleDismissUpdateSafeNotification(item, safe, dismissedNotifications, setDismissedNotifications)
+      item.onClose?.()
     },
-    [dispatch, safe, dismissedNotifications, setDismissedNotifications],
+    [dispatch],
   )
 
   // Close previous notifications in the same group
